feat: Separated logic into models
