--- Rubi/RubiModel.swift
@@ -1,18 +1,0 @@
-//
-//  RubiModel.swift
-//  Rubi
-//
-//  Created by 佐川晴海 on 2019/06/26.
-//  Copyright © 2019 佐川晴海. All rights reserved.
-//
-
-import Foundation
-
-protocol RubiModel: class{
-    
-}
-
-
-class RubiModelImpl: RubiModel {
-    
-}
