+++ Rubi/View/RubiViewController.swift
@@ -1,0 +1,28 @@
+//
+//  RubiViewController.swift
+//  Rubi
+//
+//  Created by 佐川晴海 on 2019/06/26.
+//  Copyright © 2019 佐川晴海. All rights reserved.
+//
+
+import UIKit
+
+class RubiViewController: UIViewController {
+    
+    lazy var presenter: RubiPresenter =  {
+        return RubiPresenterImpl(model: RubiModelImpl(), output: self)
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.view.backgroundColor = UIColor(red: 247/255, green: 247/255, blue: 247/255, alpha: 1)
+        presenter.requestAPI(text: "百鬼夜行")
+    }
+}
+
+extension RubiViewController: RubiPresenterOutput {
+    func convertText(hiragana: String) {
+        print(hiragana)
+    }
+}
